# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hub Comercial iSolutis — API",
    "version": "2.0.0"
  },
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Autenticação"
        ],
        "summary": "Login",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginEntrada"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenSaida"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/eu": {
      "get": {
        "tags": [
          "Autenticação"
        ],
        "summary": "Eu",
        "operationId": "eu_api_v1_auth_eu_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioLeitura"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/trocar-senha": {
      "post": {
        "tags": [
          "Autenticação"
        ],
        "summary": "Trocar Senha",
        "operationId": "trocar_senha_api_v1_auth_trocar_senha_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrocarSenhaEntrada"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/equipe": {
      "get": {
        "tags": [
          "Equipe"
        ],
        "summary": "Equipe",
        "description": "Pessoas da equipe (qualquer usuário logado): usado em responsáveis e na autoria dos registros.",
        "operationId": "equipe_api_v1_equipe_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MembroEquipe"
                  },
                  "type": "array",
                  "title": "Response Equipe Api V1 Equipe Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/usuarios": {
      "get": {
        "tags": [
          "Equipe",
          "Administração"
        ],
        "summary": "Listar Usuarios",
        "operationId": "listar_usuarios_api_v1_usuarios_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UsuarioLeitura"
                  },
                  "type": "array",
                  "title": "Response Listar Usuarios Api V1 Usuarios Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Equipe",
          "Administração"
        ],
        "summary": "Criar Usuario",
        "operationId": "criar_usuario_api_v1_usuarios_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioCriar"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioLeitura"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/usuarios/{id_}": {
      "put": {
        "tags": [
          "Equipe",
          "Administração"
        ],
        "summary": "Atualizar Usuario",
        "operationId": "atualizar_usuario_api_v1_usuarios__id___put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id "
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioAtualizar"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioLeitura"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Equipe",
          "Administração"
        ],
        "summary": "Remover Usuario",
        "operationId": "remover_usuario_api_v1_usuarios__id___delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id "
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clientes": {
      "get": {
        "tags": [
          "Clientes"
        ],
        "summary": "Listar",
        "operationId": "listar_api_v1_clientes_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ClienteResumo"
                  },
                  "type": "array",
                  "title": "Response Listar Api V1 Clientes Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Clientes"
        ],
        "summary": "Criar",
        "operationId": "criar_api_v1_clientes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteEntrada"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteLeitura"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/clientes/{id_}": {
      "put": {
        "tags": [
          "Clientes"
        ],
        "summary": "Atualizar",
        "operationId": "atualizar_api_v1_clientes__id___put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id "
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteAtualizar"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteLeitura"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Clientes"
        ],
        "summary": "Excluir",
        "operationId": "excluir_api_v1_clientes__id___delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id "
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clientes/{id_}/relacionados": {
      "get": {
        "tags": [
          "Clientes"
        ],
        "summary": "Relacionados",
        "operationId": "relacionados_api_v1_clientes__id___relacionados_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id "
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteRelacionados"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClienteEntrada": {
        "properties": {
          "nome": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Nome"
          },
          "cnpj": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cnpj"
          },
          "segmento": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Segmento"
          },
          "contato": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contato"
          },
          "cargo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cargo"
          },
          "telefone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefone"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "cidade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cidade"
          },
          "origem": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "Site",
                  "Indicação",
                  "LinkedIn",
                  "Instagram",
                  "WhatsApp",
                  "Evento",
                  "Prospecção ativa",
                  "Outro"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Origem"
          },
          "obs": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "maxLength": 20000,
            "title": "Obs"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "nome"
        ],
        "title": "ClienteEntrada"
      },
      "ClienteLeitura": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "criado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Criado Em"
          },
          "atualizado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Atualizado Em"
          },
          "criado_por": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Criado Por"
          },
          "atualizado_por": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Atualizado Por"
          },
          "versao": {
            "type": "integer",
            "title": "Versao"
          },
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "cnpj": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cnpj"
          },
          "segmento": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Segmento"
          },
          "contato": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contato"
          },
          "cargo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cargo"
          },
          "telefone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefone"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "cidade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cidade"
          },
          "origem": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origem"
          },
          "obs": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Obs"
          }
        },
        "type": "object",
        "required": [
          "id",
          "criado_em",
          "atualizado_em",
          "criado_por",
          "atualizado_por",
          "versao",
          "nome",
          "cnpj",
          "segmento",
          "contato",
          "cargo",
          "telefone",
          "email",
          "cidade",
          "origem",
          "obs"
        ],
        "title": "ClienteLeitura"
      },
      "ClienteRelacionados": {
        "properties": {
          "negocios": {
            "items": {
              "$ref": "#/components/schemas/RelNegocio"
            },
            "type": "array",
            "title": "Negocios"
          },
          "orcamentos": {
            "items": {
              "$ref": "#/components/schemas/RelOrcamento"
            },
            "type": "array",
            "title": "Orcamentos"
          },
          "lancamentos": {
            "type": "integer",
            "title": "Lancamentos"
          },
          "recebido": {
            "type": "number",
            "title": "Recebido"
          }
        },
        "type": "object",
        "required": [
          "negocios",
          "orcamentos",
          "lancamentos",
          "recebido"
        ],
        "title": "ClienteRelacionados"
      },
      "ClienteResumo": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "criado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Criado Em"
          },
          "atualizado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Atualizado Em"
          },
          "criado_por": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Criado Por"
          },
          "atualizado_por": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Atualizado Por"
          },
          "versao": {
            "type": "integer",
            "title": "Versao"
          },
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "cnpj": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cnpj"
          },
          "segmento": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Segmento"
          },
          "contato": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contato"
          },
          "cargo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cargo"
          },
          "telefone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefone"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "cidade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cidade"
          },
          "origem": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origem"
          },
          "obs": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Obs"
          },
          "negocios_abertos": {
            "type": "integer",
            "title": "Negocios Abertos"
          },
          "faturado": {
            "type": "number",
            "title": "Faturado"
          }
        },
        "type": "object",
        "required": [
          "id",
          "criado_em",
          "atualizado_em",
          "criado_por",
          "atualizado_por",
          "versao",
          "nome",
          "cnpj",
          "segmento",
          "contato",
          "cargo",
          "telefone",
          "email",
          "cidade",
          "origem",
          "obs",
          "negocios_abertos",
          "faturado"
        ],
        "title": "ClienteResumo",
        "description": "Linha da lista de clientes, com os números que a tabela mostra."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginEntrada": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "senha": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Senha"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "senha"
        ],
        "title": "LoginEntrada"
      },
      "MembroEquipe": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "ativo": {
            "type": "boolean",
            "title": "Ativo"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nome",
          "email",
          "ativo"
        ],
        "title": "MembroEquipe",
        "description": "Visão mínima da equipe, disponível a qualquer pessoa logada (responsáveis, autoria)."
      },
...
      "RelOrcamento": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "numero": {
            "type": "string",
            "title": "Numero"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "total_projeto": {
            "type": "number",
            "title": "Total Projeto"
          }
        },
        "type": "object",
        "required": [
          "id",
          "numero",
          "status",
          "total_projeto"
        ],
        "title": "RelOrcamento"
      },
...
      "TokenSaida": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "usuario": {
            "$ref": "#/components/schemas/UsuarioLeitura"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "usuario"
        ],
        "title": "TokenSaida"
      },
      "TrocarSenhaEntrada": {
        "properties": {
          "senha_atual": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Senha Atual"
          },
          "nova_senha": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Nova Senha"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "senha_atual",
          "nova_senha"
        ],
        "title": "TrocarSenhaEntrada"
      },
      "UsuarioAtualizar": {
        "properties": {
          "versao": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Versao"
          },
          "nome": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Nome"
          },
          "admin": {
            "type": "boolean",
            "title": "Admin"
          },
          "ativo": {
            "type": "boolean",
            "title": "Ativo",
            "default": true
          },
          "senha": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 8
              },
              {
                "type": "null"
              }
            ],
            "title": "Senha"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "versao",
          "nome",
          "admin"
        ],
        "title": "UsuarioAtualizar"
      },
      "UsuarioCriar": {
        "properties": {
          "nome": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Nome"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "senha": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Senha"
          },
          "admin": {
            "type": "boolean",
            "title": "Admin",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "nome",
          "email",
          "senha"
        ],
        "title": "UsuarioCriar"
      },
      "UsuarioLeitura": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "admin": {
            "type": "boolean",
            "title": "Admin"
          },
          "ativo": {
            "type": "boolean",
            "title": "Ativo"
          },
          "senha_definida": {
            "type": "boolean",
            "title": "Senha Definida"
          },
          "ultimo_acesso": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ultimo Acesso"
          },
          "versao": {
            "type": "integer",
            "title": "Versao"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "nome",
          "admin",
          "ativo",
          "senha_definida",
          "ultimo_acesso",
          "versao"
        ],
        "title": "UsuarioLeitura"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}